Fix naming of mappers
